requirements.txt
# Asosiy
Django==5.2.1
djangorestframework==3.16.1
psycopg2==2.9.9
gunicorn==21.2.0

# Auth
djoser==2.3.3
djangorestframework_simplejwt==5.5.1
social-auth-app-django==5.7.0
social-auth-core==4.8.3
oauthlib==3.3.1
requests-oauthlib==2.0.0
python3-openid==3.2.0
defusedxml==0.7.1
PyJWT==2.11.0
cryptography==46.0.4

# CORS & Filter
django-cors-headers==4.9.0
django-filter==25.2

# Database & tasks
celery==5.6.2
redis==7.2.0
pillow==12.1.1
kombu==5.6.2
vine==5.1.0
billiard==4.2.4
amqp==5.3.1
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
prompt_toolkit==3.0.52
wcwidth==0.6.0

# Bot
python-telegram-bot==22.6
httpx==0.28.1
httpcore==1.0.9
h11==0.16.0
certifi==2026.1.4
idna==3.11
sniffio==1.3.0  # (anyio talab qiladi, ro‘yxatda yo‘q edi)
anyio==4.12.1

# HTTP requests
requests==2.32.5
urllib3==2.6.3
charset-normalizer==3.4.4

# Django utilities
asgiref==3.11.0
sqlparse==0.5.5
tzdata==2025.3
pytz==2025.2
python-dateutil==2.9.0.post0
six==1.17.0

# Environment
python-dotenv==1.2.2

# Docs (agar kerak bo‘lsa)
drf-yasg==1.21.14
PyYAML==6.0.3
uritemplate==4.2.0
inflection==0.5.1
packaging==26.0
